chore: update logger dependency to version ^2.5.0 and refactor import paths for better organization in API Explorer
