+++ projects/api_explorer/lib/widgets/home/modern_api_panel.dart
@@ -1,9 +1,9 @@
-import 'package:api_explorer/styles/app_theme.dart';
-import 'package:api_explorer/widgets/beautiful_gradient_button.dart';
-import 'package:api_explorer/widgets/http_method_selector.dart';
-import 'package:api_explorer/widgets/panel_header.dart';
-import 'package:api_explorer/widgets/service_info.dart';
 import 'package:flutter/material.dart';
+import 'package:api_explorer/styles/app_theme.dart';
+import 'package:api_explorer/widgets/common/beautiful_gradient_button.dart';
+import 'package:api_explorer/widgets/home/http_method_selector.dart';
+import 'package:api_explorer/widgets/home/panel_header.dart';
+import 'package:api_explorer/widgets/home/service_info.dart';
 import 'package:api_explorer/services/api_service_registry.dart';
 
 class ModernApiPanel extends StatefulWidget {
