+++ projects/api_explorer/lib/widgets/layout/desktop_layout_widget.dart
@@ -1,6 +1,6 @@
-import 'package:api_explorer/widgets/ide_response_panel.dart';
-import 'package:api_explorer/widgets/modern_api_panel.dart';
-import 'package:api_explorer/widgets/modern_sidebar.dart';
+import 'package:api_explorer/widgets/home/ide_response_panel.dart';
+import 'package:api_explorer/widgets/home/modern_api_panel.dart';
+import 'package:api_explorer/widgets/home/modern_sidebar.dart';
 import 'package:flutter/material.dart';
 import 'package:api_explorer/services/api_service_registry.dart';
