+++ projects/api_explorer/lib/widgets/layout/mobile_app_bar_widget.dart
