+++ projects/api_explorer/lib/widgets/layout/mobile_layout_widget.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:api_explorer/services/api_service_registry.dart';
-import 'package:api_explorer/widgets/ide_response_panel.dart';
-import 'package:api_explorer/widgets/modern_api_panel.dart';
-import 'package:api_explorer/widgets/api_explorer/mobile_app_bar_widget.dart';
-import 'package:api_explorer/widgets/api_explorer/mobile_bottom_nav_widget.dart';
+import 'package:api_explorer/widgets/home/ide_response_panel.dart';
+import 'package:api_explorer/widgets/home/modern_api_panel.dart';
+import 'package:api_explorer/widgets/layout/mobile_app_bar_widget.dart';
+import 'package:api_explorer/widgets/layout/mobile_bottom_nav_widget.dart';
 
 class MobileLayoutWidget extends StatelessWidget {
   final BoxConstraints constraints;
