+++ projects/api_explorer/lib/widgets/store_management/store_setup_wizard.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:apis/apis.dart';
 import 'package:get_it/get_it.dart';
